+++ evo1/seq/crystal_cavern.py
@@ -1,13 +1,19 @@
 import logging
 
 import evo1.control
-from engine.mathlib import Facing, Vec2, is_close
+from engine.mathlib import Facing, Vec2
 from engine.seq import SeqList
-from evo1.atb import EncounterID, SeqATBCombat, SeqATBmove2D, calc_next_encounter
+from evo1.atb import EncounterID, SeqATBCombat, SeqATBmove2D
+from evo1.interact import SeqMashDelay
 from evo1.maps import GetAStar
-from evo1.memory import MapID, get_memory, get_zelda_memory
-from evo1.move2d import SeqGrabChest, SeqHoldInPlace, SeqMove2D, SeqZoneTransition
-from memory.rng import EvolandRNG
+from evo1.memory import MapID, get_zelda_memory
+from evo1.move2d import (
+    SeqGrabChest,
+    SeqHoldInPlace,
+    SeqMove2D,
+    SeqMove2DConfirm,
+    SeqZoneTransition,
+)
 from term.window import WindowLayout
 
 logger = logging.getLogger(__name__)
@@ -34,7 +40,7 @@ class CrystalCavernEncManip(SeqATBmove2D):
         enc_timer = mem.player.encounter_timer
 
         # If we are about to get an encounter, potentially manip (stop and wait)
-        if enc_timer < 0.2:
+        if enc_timer < 0.1:
             if self.next_enc.enc_id in self.pref_enc:
                 # If favorable and it's a Kobra, wait if we're going to miss
                 return (
@@ -49,12 +55,7 @@ class CrystalCavernEncManip(SeqATBmove2D):
 
     # Returning true means we seize control instead of moving on
     def do_encounter_manip(self) -> bool:
-        mem = get_memory()
-        rng = EvolandRNG().get_rng()
-        self.next_enc = calc_next_encounter(
-            rng=rng, has_3d_monsters=False, clink_level=mem.lvl
-        )
-
+        self.calc_next_encounter()
         if not self._should_manip():
             return False
 
@@ -68,12 +69,9 @@ class CrystalCavernEncManip(SeqATBmove2D):
         super().render(window=window)
 
 
-# Dummy class for ATB combat testing; requires manual control
 class SeqKefkasGhost(SeqATBCombat):
-    def __init__(self, target: Vec2) -> None:
+    def __init__(self) -> None:
         super().__init__(name="Kefka's Ghost")
-        self.target = target
-        self.precision = 0.2
 
     # Kefka's Ghost has a strange property; when the boss turns invincible,
     # its turn gauge will be set to -999999999999. We should avoid attacking
@@ -82,6 +80,9 @@ class SeqKefkasGhost(SeqATBCombat):
         return self.mem.enemies[0].turn_gauge < -1
 
     def handle_combat(self):
+        # Do nothing if battle has ended
+        if self.mem.ended:
+            return
         # Do nothing if the boss is in the counter phase
         if self._is_invincible():
             return
@@ -92,10 +93,7 @@ class SeqKefkasGhost(SeqATBCombat):
 
     def execute(self, delta: float) -> bool:
         super().execute(delta)
-        # Check if we have reached the goal
-        mem = get_zelda_memory()
-        player_pos = mem.player.pos
-        return is_close(player_pos, self.target, precision=self.precision)
+        return self.state == self._BattleFSM.POST_BATTLE
 
 
 class CrystalCavern(SeqList):
@@ -103,25 +101,37 @@ class CrystalCavern(SeqList):
         super().__init__(
             name="Crystal Cavern",
             children=[
+                SeqMove2DConfirm(
+                    name="Move to chest",
+                    coords=_cavern_astar.calculate(
+                        start=Vec2(24, 77), goal=Vec2(20, 66)
+                    ),
+                ),
+                SeqMove2D("Move to chest", coords=[Vec2(20, 65.6)]),
+                SeqGrabChest("Cave monsters", Facing.UP),
+                # TODO: Should run from these battles
                 CrystalCavernEncManip(
                     name="Move to chest",
                     coords=_cavern_astar.calculate(
-                        start=Vec2(24, 77), goal=Vec2(18, 39)
+                        start=Vec2(20, 65), goal=Vec2(18, 39)
                     ),
                     pref_enc=[EncounterID.KOBRA],
                 ),
                 SeqGrabChest("Experience", Facing.UP),
+                # TODO: Allowed battles may be too tight
+                # TODO: Aim is to get level 2, so need Tork or 2xKobra (pref not Tork+2xScaven)
                 CrystalCavernEncManip(
                     name="Move to trigger",
                     coords=_cavern_astar.calculate(
                         start=Vec2(18, 38), goal=Vec2(54, 36)
                     ),
-                    pref_enc=[EncounterID.TORK],
+                    pref_enc=[EncounterID.TORK, EncounterID.KOBRA_2],
                 ),
                 # TODO should menu manip here
                 SeqHoldInPlace(
                     name="Trigger plate", target=Vec2(54, 36.5), timeout_in_s=0.5
                 ),
+                # TODO: Allowed battles may be too tight
                 CrystalCavernEncManip(
                     name="Move to boss",
                     coords=_cavern_astar.calculate(
@@ -130,7 +140,9 @@ class CrystalCavern(SeqList):
                     pref_enc=[EncounterID.KOBRA],
                 ),
                 # Trigger fight against Kefka's ghost (interact with crystal)
-                SeqKefkasGhost(target=Vec2(7, 10)),
+                SeqKefkasGhost(),
+                # Mash past end of battle, activate the crystal but don't talk to Kaeris
+                SeqMashDelay("Activate crystal", timeout_in_s=3.0),
                 # Limbo realm
                 SeqMove2D(
                     name="Move to portal",